add makeflow support for specifying input deps

+++ makeflow/src/makeflow.c
@@ -9,6 +9,7 @@ See the file COPYING for details.
 #include "batch_job.h"
 #include "cctools.h"
 #include "copy_stream.h"
+#include "create_dir.h"
 #include "debug.h"
 #include "getopt_aux.h"
 #include "hash_table.h"
@@ -38,6 +39,7 @@ See the file COPYING for details.
 #include "makeflow_wrapper_docker.h"
 #include "makeflow_wrapper_monitor.h"
 #include "makeflow_wrapper_umbrella.h"
+#include "makeflow_mounts.h"
 #include "parser.h"
 #include "parser_jx.h"
 
@@ -46,6 +48,7 @@ See the file COPYING for details.
 #include <fcntl.h>
 #include <sys/stat.h>
 #include <sys/types.h>
+#include <libgen.h>
 
 #include <errno.h>
 #include <signal.h>
@@ -141,6 +144,10 @@ static struct makeflow_wrapper_umbrella *wrapper_umbrella = 0;
 /* is true only if the -N command is used. Useful for catalog reporting.
  */
 static int catalog_reporting_on = 0;
+
+static char *mountfile = NULL;
+static char *mount_cache = NULL;
+static int use_mountfile = 0;
 
 /* Generates file list for node based on node files, wrapper
  *  * input files, and monitor input files. Relies on %% nodeid
@@ -813,6 +820,11 @@ static int makeflow_check(struct dag *d)
 			if(skip_file_check || batch_fs_stat(remote_queue, f->filename, &buf) >= 0) {
 				continue;
 			}
+
+			if(f->source) {
+				continue;
+			}
+
 			fprintf(stderr, "makeflow: %s does not exist, and is not created by any rule.\n", f->filename);
 			error++;
 		}
@@ -1001,7 +1013,7 @@ static void show_help_run(const char *cmd)
 {
 	printf("Use: %s [options] <dagfile>\n", cmd);
 	printf("Frequently used options:\n\n");
-	printf(" %-30s Clean up: remove logfile and all targets. Optional specification [intermediates, outputs] removes only the indicated files.\n", "-c,--clean=<type>");
+	printf(" %-30s Clean up: remove logfile and all targets. Optional specification [intermediates, outputs, cache] removes only the indicated files.\n", "-c,--clean=<type>");
 	printf(" %-30s Batch system type: (default is local)\n", "-T,--batch-type=<type>");
 	printf(" %-30s %s\n\n", "", batch_queue_type_string());
 	printf("Other options are:\n");
@@ -1021,6 +1033,8 @@ static void show_help_run(const char *cmd)
 	printf(" %-30s Use this file for the makeflow log.		 (default is X.makeflowlog)\n", "-l,--makeflow-log=<logfile>");
 	printf(" %-30s Use this file for the batch system log.	 (default is X.<type>log)\n", "-L,--batch-log=<logfile>");
 	printf(" %-30s Send summary of workflow to this email address upon success or failure.\n", "-m,--email=<email>");
+	printf(" %-30s Use this file as a mountlist.\n", "   --mounts=<mountfile>");
+	printf(" %-30s Use this dir as the cache for file dependencies.\n", "   --cache=<cache_dir>");
 	printf(" %-30s Set the project name to <project>\n", "-N,--project-name=<project>");
 	printf(" %-30s Send debugging to this file. (can also be :stderr, :stdout, :syslog, or :journal)\n", "-o,--debug-file=<file>");
 	printf(" %-30s Rotate debug file once it reaches this size.\n", "   --debug-rotate-max=<bytes>");
@@ -1125,6 +1139,7 @@ int main(int argc, char *argv[])
 
 	enum {
 		LONG_OPT_AUTH = UCHAR_MAX+1,
+		LONG_OPT_CACHE,
 		LONG_OPT_DEBUG_ROTATE_MAX,
 		LONG_OPT_DISABLE_BATCH_CACHE,
 		LONG_OPT_DOT_CONDENSE,
@@ -1135,6 +1150,7 @@ int main(int argc, char *argv[])
 		LONG_OPT_MONITOR_LOG_NAME,
 		LONG_OPT_MONITOR_OPENED_FILES,
 		LONG_OPT_MONITOR_TIME_SERIES,
+		LONG_OPT_MOUNTS,
 		LONG_OPT_PASSWORD,
 		LONG_OPT_TICKETS,
 		LONG_OPT_VERBOSE_PARSING,
@@ -1163,6 +1179,7 @@ int main(int argc, char *argv[])
 		{"batch-log", required_argument, 0, 'L'},
 		{"batch-options", required_argument, 0, 'B'},
 		{"batch-type", required_argument, 0, 'T'},
+		{"cache", required_argument, 0, LONG_OPT_CACHE},
 		{"catalog-server", required_argument, 0, 'C'},
 		{"clean", optional_argument, 0, 'c'},
 		{"debug", required_argument, 0, 'd'},
@@ -1185,6 +1202,7 @@ int main(int argc, char *argv[])
 		{"monitor-log-name", required_argument, 0, LONG_OPT_MONITOR_LOG_NAME},
 		{"monitor-with-opened-files", no_argument, 0, LONG_OPT_MONITOR_OPENED_FILES},
 		{"monitor-with-time-series",  no_argument, 0, LONG_OPT_MONITOR_TIME_SERIES},
+		{"mounts",  required_argument, 0, LONG_OPT_MOUNTS},
 		{"password", required_argument, 0, LONG_OPT_PASSWORD},
 		{"port", required_argument, 0, 'p'},
 		{"port-file", required_argument, 0, 'Z'},
@@ -1242,6 +1260,8 @@ int main(int argc, char *argv[])
 						clean_mode = MAKEFLOW_CLEAN_INTERMEDIATES;
 					} else if(strcasecmp(optarg, "outputs") == 0){
 						clean_mode = MAKEFLOW_CLEAN_OUTPUTS;
+					} else if(strcasecmp(optarg, "cache") == 0){
+						clean_mode = MAKEFLOW_CLEAN_CACHE;
 					} else if(strcasecmp(optarg, "all") != 0){
 						fprintf(stderr, "makeflow: unknown clean option %s", optarg);
 						exit(1);
@@ -1339,6 +1359,12 @@ int main(int argc, char *argv[])
 				if (!monitor) monitor = makeflow_monitor_create();
 				if(log_format) free(log_format);
 				log_format = xxstrdup(optarg);
+				break;
+			case LONG_OPT_CACHE:
+				mount_cache = xxstrdup(optarg);
+				break;
+			case LONG_OPT_MOUNTS:
+				mountfile = xxstrdup(optarg);
 				break;
 			case LONG_OPT_AMAZON_CREDENTIALS:
 				amazon_credentials = xxstrdup(optarg);
@@ -1674,6 +1700,19 @@ int main(int argc, char *argv[])
 	if (change_dir)
 		chdir(change_dir);
 
+	/* Prepare the input files specified in the mountfile. */
+	if(mountfile && !clean_mode) {
+		/* check the validity of the mountfile and load the info from the mountfile into the dag */
+		printf("checking the consistency of the mountfile ...\n");
+		if(makeflow_mounts_parse_mountfile(mountfile, d)) {
+			fprintf(stderr, "Failed to parse the mountfile: %s.\n", mountfile);
+			free(mountfile);
+			return -1;
+		}
+		free(mountfile);
+		use_mountfile = 1;
+	}
+
 	printf("checking %s for consistency...\n",dagfile);
 	if(!makeflow_check(d)) {
 		exit(EXIT_FAILURE);
@@ -1687,7 +1726,33 @@ int main(int argc, char *argv[])
 	setlinebuf(stdout);
 	setlinebuf(stderr);
 
-	makeflow_log_recover(d, logfilename, log_verbose_mode, remote_queue, clean_mode, skip_file_check );
+	if(mount_cache) d->cache_dir = mount_cache;
+
+	/* In case when the user uses --cache option to specify the mount cache dir and the log file also has
+	 * a cache dir logged, these two dirs must be the same. Otherwise exit.
+	 */
+	if(makeflow_log_recover(d, logfilename, log_verbose_mode, remote_queue, clean_mode, skip_file_check )) {
+		dag_mount_clean(d);
+		exit(EXIT_FAILURE);
+	}
+
+	/* This check must happen after makeflow_log_recover which may load the cache_dir info into d->cache_dir.
+	 * This check must happen before makeflow_mount_install to guarantee that the program ends before any mount is copied if any target is invliad.
+	 */
+	if(use_mountfile) {
+		if(makeflow_mount_check_target(d)) {
+			dag_mount_clean(d);
+			exit(EXIT_FAILURE);
+		}
+	}
+
+	if(use_mountfile && !clean_mode) {
+		if(makeflow_mounts_install(d)) {
+			fprintf(stderr, "Failed to install the dependencies specified in the mountfile!\n");
+			dag_mount_clean(d);
+			exit(EXIT_FAILURE);
+		}
+	}
 
 	if(monitor) {
 		if(!log_dir)
@@ -1715,13 +1780,22 @@ int main(int argc, char *argv[])
 
 	if(clean_mode != MAKEFLOW_CLEAN_NONE) {
 		printf("cleaning filesystem...\n");
-		makeflow_clean(d, remote_queue, clean_mode);
+		if(makeflow_clean(d, remote_queue, clean_mode)) {
+			fprintf(stderr, "Failed to clean up makeflow!\n");
+			exit(EXIT_FAILURE);
+		}
+
 		if(clean_mode == MAKEFLOW_CLEAN_ALL) {
 			unlink(logfilename);
 		}
+
 		exit(0);
 	}
 
+	/* this func call guarantees the mount fields set up from the info of the makeflow log file are cleaned up
+	 * even if the user does not use --mounts or -c option.
+	 */
+	dag_mount_clean(d);
 
 	printf("starting workflow....\n");
